+++ src/Application/FormType/CustomerType.php
@@ -7,7 +7,6 @@ namespace Iriven\PhpFormGenerator\Application\FormType;
 use Iriven\PhpFormGenerator\Domain\Constraint\Required;
 use Iriven\PhpFormGenerator\Domain\Contract\FormTypeInterface;
 use Iriven\PhpFormGenerator\Domain\Contract\OptionsResolverInterface;
-use Iriven\PhpFormGenerator\Domain\Field\CountryType;
 use Iriven\PhpFormGenerator\Domain\Field\EmailType;
 use Iriven\PhpFormGenerator\Domain\Field\TextType;
 use Iriven\PhpFormGenerator\Domain\Form\FormBuilder;
@@ -18,20 +17,14 @@ final class CustomerType implements FormTypeInterface
     public function buildForm(FormBuilder $builder, array $options = []): void
     {
         $builder
-            ->add('name', TextType::class, [
-                'label' => 'Customer name',
-                'constraints' => [new Required()],
-            ])
-            ->add('email', EmailType::class, [
-                'label' => 'Customer email',
-            ])
-            ->add('country', CountryType::class, [
-                'label' => 'Country',
-            ]);
+            ->add('name', TextType::class, ['constraints' => [new Required()]])
+            ->add('email', EmailType::class, ['constraints' => [new Required()]]);
     }
 
     public function configureOptions(OptionsResolverInterface $resolver): void
     {
-        $resolver->setDefaults([]);
+        $resolver->setDefaults([
+            'csrf_protection' => true,
+        ]);
     }
 }